+++ app/src/processing/app/api/Sketchbook.kt
@@ -3,10 +3,16 @@ package processing.app.api
 import com.github.ajalt.clikt.command.SuspendingCliktCommand
 import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.core.subcommands
-import processing.app.Base
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import processing.app.Platform
+import processing.app.Preferences
 import processing.app.api.Sketch.Companion.getSketches
+import java.io.File
 
 class Sketchbook: SuspendingCliktCommand() {
+
+
     override fun help(context: Context) = "Manage the sketchbook"
     override suspend fun run() {
         System.setProperty("java.awt.headless", "true")
@@ -14,12 +20,23 @@ class Sketchbook: SuspendingCliktCommand() {
     init {
         subcommands(SketchbookList())
     }
+
+
     class SketchbookList: SuspendingCliktCommand("list") {
+        val serializer = Json {
+            prettyPrint = true
+        }
+
         override fun help(context: Context) = "List all sketches"
         override suspend fun run() {
-            val sketchbookFolder = Base.getSketchbookFolder()
+            Platform.init()
+            // TODO: Allow the user to change the sketchbook location
+            // TODO: Currently blocked since `Base.getSketchbookFolder()` is not available in headless mode
+            val sketchbookFolder = Platform.getDefaultSketchbookFolder()
 
-            val sketches = getSketches(sketchbookFolder)
+            val sketches = getSketches(sketchbookFolder.resolve("sketchbook"))
+            val json = serializer.encodeToString(listOf(sketches))
+            println(json)
         }
     }
 }
